doc-gen: extract RPC helpers into rpc.go
Pure refactor: no behavior change. Splits the RPC-specific code
(BITCOIN_*, Command/Group, getVersion, generateRPC, run) out of
generate.go into rpc.go, leaving generate.go as a thin entry point.

Adds a go.mod so multi-file builds work with `go run .`.
