+++ contrib/doc-gen/go.mod
@@ -1,0 +1,3 @@
+module bitcoincore.org/contrib/doc-gen
+
+go 1.21
